requirements.txt
prefect==2.13.1
prefect-aws==0.3.7
s3fs==2023.5.0
requests==2.30.0
flake8==6.0.0
pyarrow==12.0.0
python-dotenv==1.0.0
subgrounds==1.6.0
web3==6.4.0
pycoingecko==3.1.0
pycountry==22.3.5
backoff==2.2.1
